+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Pre-release]
+### Added
+- an option to specify CSS selectors that would be included in changes discovery along with the existing `h-entry`
+- an option to specify CSS selectors that would be excluded from gathering targets when sending webmentions
 
 ## [0.4.0] - 2021-10-03
 ### Changed